pyproject.toml
[tool.poetry]
name = "amili-scraper"
version = "0.1.0"
description = "Scrapes the web for nutritional data."
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
python-dotenv = "^0.15.0"
msedge-selenium-tools = "^3.141.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
